refactor: Introduce a convenience class for managing the cli
This changes borrows the cli convenience classes from build-tools and uses them to slim down the amount of command line management in the main changes script. In the process, it was necessary to change the way the scope was handled to ensure the legacy behaviour continued to work alongside the new per-command definitions.