refactor: Introduce a convenience class for managing the cli (#29)
This changes borrows the cli convenience classes from build-tools and uses them to slim down the amount of command line management in the main changes script. In the process, it was necessary to modify the way the scope was handled to ensure the legacy behaviour continued to work alongside the new per-command arguments.

This includes a drive-by fix to improve stderr logging for test debugging by enabling debug output if the `DEBUG` environment variable equals `1`.