feat: Customisable templates for rendering release notes (#47)
This change switches to Jinja2 for rendering the release notes instead of hand-crafted string manipulation. It introduces two default templates (one for single release release notes, and one for multiple release release notes). It also introduces a parameter to the `notes` command that allows users to override these templates and provide their own.