feat: Pass CHANGES_PRERELEASE environment variable to release commands (#50)
Some release tools (looking at you GitHub) require releases to be explicitly marked as prerelease. This change introduces an additional environment variable (`CHANGES_PRERELEASE`) which can be inspected to determine if the version is prerelease or not (major version 0). It also introduces unit tests and some explicit tests for the shell being used to execute commands.