fix!: Rename `CHANGES_PRERELEASE` to `CHANGES_INITIAL_DEVELOPMENT` (#75)
This is a preparatory change in advance of adding support for SemVer pre-release versioning. Since--under-the-hood--the `CHANGES_PRERELEASE` environment variable was checking to see if the major version is greater than 0, it has been renamed to 'initial development' correspond with SemVer's conventions (see https://semver.org/#spec-item-4). 'Unstable' was considered, but this seems less explicit and might not correctly represent all projects in the pre-1.0.0 state.